Fix : 대기 씬 스크립트 수정
플레이어 엔트리 : 플레이어 생성 코드 삭제
룸 매니저 : 플레이어 생성 코드 추가
룸 패널 : 플레이어 생성 코드 삭제
