+++ Assets/Works/MKH/MKH_WaitingRoom/MKH_Scripts/MKH_RoomPanel.cs
@@ -1,27 +1,21 @@
 using ExitGames.Client.Photon;
 using Photon.Pun;
+using Photon.Pun.Demo.PunBasics;
 using Photon.Pun.UtilityScripts;
 using Photon.Realtime;
 using UnityEngine;
 using UnityEngine.UI;
 
 
-public class MKH_RoomPanel : MonoBehaviour
+public class MKH_RoomPanel : MonoBehaviourPun
 {
     [SerializeField] MKH_PlayerEntry[] playerEntries;
-    [SerializeField] MKH_PlayerManager[] playerManagers;
     [SerializeField] Button startButton;
-
-    private void Start()
-    {
-        
-    }
 
     // 방에 들어왔을 때
     private void OnEnable()
     {
         UpdatePlayers();
-        PlayerSpawn();
         // 플레이어 넘버링 업데이트
         PlayerNumbering.OnPlayerNumberingChanged += UpdatePlayers;
 
@@ -38,15 +32,15 @@ public class MKH_RoomPanel : MonoBehaviour
     public void UpdatePlayers()
     {
         // 플레이어 들어오기전 공간 셋팅
-        foreach (MKH_PlayerEntry entry in playerEntries)          
+        foreach (MKH_PlayerEntry entry in playerEntries)
         {
             entry.SetEmpty();
         }
         // 모든 플레이어 확인
-        foreach (Player player in PhotonNetwork.PlayerList)        
+        foreach (Player player in PhotonNetwork.PlayerList)
         {
             // 아직 번호 할당 안받았을 때 업데이트 하지 말기
-            if (player.GetPlayerNumber() == -1)                     
+            if (player.GetPlayerNumber() == -1)
                 continue;
             // 플레이어 넘버 가지고 오기
             int number = player.GetPlayerNumber();
@@ -95,7 +89,7 @@ public class MKH_RoomPanel : MonoBehaviour
     private bool CheckAllReady()
     {
         // 플레이어 리스트에 있는 플레이어들이 레디를 안했을 경우 비활성화
-        foreach(Player player in PhotonNetwork.PlayerList)
+        foreach (Player player in PhotonNetwork.PlayerList)
         {
             if (player.GetReady() == false)
                 return false;
@@ -117,13 +111,5 @@ public class MKH_RoomPanel : MonoBehaviour
     public void LeaveRoom()
     {
         PhotonNetwork.LeaveRoom();
-    }
-
-    private void PlayerSpawn()
-    {
-        Debug.Log("1");
-        Vector3 randomPos = new Vector3(Random.Range(-5f, 5f), 0, Random.Range(-5f, 5f));
-
-        PhotonNetwork.Instantiate("GameObject/MatchMaking/Player", randomPos, Quaternion.identity);
     }
 }
